marked extra metrics tests and changed docs
