+++ .circleci/continue_config.yml
@@ -43,7 +43,7 @@ jobs:
     steps:
       - checkout
       - restore_cache:
-          key: &ci-base-cache ci-cache-<< parameters.python-version >><< parameters.extras >>-{{ checksum "pyproject.toml" }}
+          key: &ci-base-cache ci-cache-<< parameters.python-version >>-<< parameters.extras >>-{{ checksum "pyproject.toml" }}
       - when:
           condition:
             equal: [ "", << parameters.extras >> ]
@@ -56,7 +56,7 @@ jobs:
                   poetry install --no-ansi
       - when:
           condition:
-            equal: [ "-metrics", << parameters.extras >> ]
+            equal: [ "metrics", << parameters.extras >> ]
           steps:
             - run:
                 name: Install Dependencies
@@ -69,7 +69,6 @@ jobs:
           paths:
             - /home/circleci/project/.poetry/virtualenvs
             - poetry.lock
-
 
   unit-test:
     parameters:
@@ -87,7 +86,7 @@ jobs:
     steps:
       - checkout
       - restore_cache:
-          key: ci-cache-<< parameters.python-version >><< parameters.extras >>-{{ checksum "pyproject.toml" }}
+          key: ci-cache-<< parameters.python-version >>-<< parameters.extras >>-{{ checksum "pyproject.toml" }}
       - run: poetry run python3 -c 'import kolena'
       # TODO: fix underlying mypy issues with Python>3.9 rather than skipping
       - when:
@@ -105,15 +104,15 @@ jobs:
             - run:
                 name: Run unit tests
                 command: |
-                  poetry run pytest -vv --cov=kolena --cov-branch tests/unit
+                  poetry run pytest -vv --cov=kolena --cov-branch tests/unit --mark "not metrics"
       - when:
           condition:
-            equal: [ "-metrics", << parameters.extras >> ]
+            equal: [ "metrics", << parameters.extras >> ]
           steps:
             - run:
-                name: Run unit tests
+                name: Run metrics unit tests
                 command: |
-                  poetry run pytest -vv --cov=kolena --cov-branch tests/unit_extras
+                  poetry run pytest -vv --cov=kolena --cov-branch tests/unit --mark "metrics"
       - when:
           # Generate coverage only from one python version
           condition:
@@ -222,39 +221,41 @@ workflows:
   ci:
     jobs:
       - ci-base:
-          name: ci-base-<< matrix.python-version >><< matrix.extras >>
+          name: ci-base-<< matrix.python-version >>-<< matrix.extras >>
           matrix:
             parameters:
               python-version: [ "3.7", "3.8", "3.9", "3.10", "3.11" ]
-              extras: ["", "-metrics"]
+              extras: ["", "metrics"]
       - unit-test:
-          name: unit-test-<< matrix.python-version >><< matrix.extras >>
+          name: unit-test-<< matrix.python-version >>-<< matrix.extras >>
           matrix:
             parameters:
               python-version: [ "3.7", "3.8", "3.9", "3.10", "3.11" ]
-              extras: ["", "-metrics"]
+              extras: ["", "metrics"]
           requires:
-            - ci-base-<< matrix.python-version >><< matrix.extras >>
+            - ci-base-<< matrix.python-version >>-<< matrix.extras >>
       - integration-test:
-          name: integration-test-<< matrix.pytest-group >>-<< matrix.python-version >>
+          name: integration-test-<< matrix.pytest-group >>-<< matrix.python-version >>-<< matrix.extras >>
           matrix:
             parameters:
               python-version: [ "3.9" ]
               pytest-group: [ detection, fr, classification, workflow ]
+              extras: ["", "metrics"]
           enabled: true
           parallelism: 4
           requires:
-            - ci-base-<< matrix.python-version >>
+            - ci-base-<< matrix.python-version >>-<< matrix.extras >>
       - integration-test:
-          name: integration-test-misc-<< matrix.python-version >>
+          name: integration-test-misc-<< matrix.python-version >>-<< matrix.extras >>
           matrix:
             parameters:
               python-version: [ "3.9" ]
+              extras: ["", "metrics"]
           pytest-group: misc
           enabled: << pipeline.parameters.all >>
           parallelism: 1
           requires:
-            - ci-base-<< matrix.python-version >>
+            - ci-base-<< matrix.python-version >>-<< matrix.extras >>
   example:
     when:
       or: [ << pipeline.parameters.workflow >>, << pipeline.parameters.all >>, << pipeline.parameters.example >> ]
